move all .svelte under components; some cleanup
