--- src/js/App.svelte
@@ -1,8 +1,0 @@
-
-<script>
-    import Button from "./Button.svelte";
-</script>
-
-<div>
-    <Button />
-</div>
