--- src/js/Button.svelte
@@ -1,7 +1,0 @@
-<script>
-    let roger = "Roger";
-</script>
-
-<div class="mt-3">
-    <button class="btn btn-primary">Hi, {roger}!</button>
-</div>
