--- src/js/svelte.js
@@ -1,8 +1,0 @@
-import App from "./App.svelte";
-// import "bootstrap/dist/css/bootstrap.min.css";
-
-const app = new App({
-  target: document.getElementById("app"),
-});
-
-export default app;
